moved cell class to separete file
