+++ public/scripts/picture&components.js
@@ -1,7 +1,4 @@
-const State = {
-  ALIVE: 'alive',
-  DEAD: 'dead'
-};
+import { Cell } from './cell.js';
 
 class Picture {
   constructor(width, height, pallete, pixels) {
@@ -18,52 +15,6 @@ class Picture {
   copy() {
     return new Picture(this.width,
       this.height, this.pallete, this.pixels.map((cell) => cell.copy()));
-  }
-}
-
-class Cell {
-  constructor(state = State.DEAD, palette) {
-    this.state = state;
-    this.palette = palette;
-  }
-
-  get state() {
-    return this._state;
-  }
-
-  set state(value) {
-    if (value === State.DEAD || value === 0) {
-      this._state = State.DEAD;
-      this.age = 0;
-    }
-    if (value === State.ALIVE || value === 1) {
-      this._state = State.ALIVE;
-      this.age = 1;
-    }
-  }
-
-  isAlive() {
-    if (this.state === State.DEAD) return 0;
-    else return 1;
-  }
-
-  getOld() {
-    if (this._state === State.ALIVE) {
-      this.age++;
-    }
-    // add smth for else?
-  }
-
-  color() {
-    const arrLen = this.palette.length;
-    if (this.age < arrLen) return this.palette[this.age];
-    return this.palette[arrLen - 1];
-  }
-
-  copy() {
-    const newCell = new Cell(this._state, this.palette);
-    newCell.age = this.age;
-    return newCell;
   }
 }
 
@@ -90,8 +41,6 @@ function randomCells(size) {
 }
 
 export { Picture };
-export { Cell };
 export { fillPicture };
 export { deadCells };
 export { randomCells };
-export { State };
